opportunity to create advertising

+++ auth_core/models.py
@@ -3,6 +3,6 @@ from django.contrib.auth.models import AbstractUser
 
 # Create your models here.
 class User(AbstractUser):
-  avatar = models.ImageField(upload_to='users', blank=True)
-  phone = models.PositiveSmallIntegerField(null=True, blank=True)
-  location = models.TextField(blank=True)
+    avatar = models.ImageField(upload_to='users', blank=True)
+    phone = models.PositiveIntegerField(null=True, blank=True)
+    location = models.TextField(blank=True)
